+++ CHANGELOG.md
@@ -54,6 +54,8 @@ For instructions on upgrading to newer versions, visit
 
 ## 2.4.2 \[ In Development \] \[ Branch: 2.4.0-stable \]
 
+* \#1605 Remove deprecation warning on rescue responses, Rails 3.2
+
 * \#1602 Preserve structure of $and and $or queries when typecasting.
 
 * \#1600 Uniqueness validation no longer errors when provided a relation.